- encapsulate all the stuff related to process managers configuration in a single method
