+++ src/Orchestration/NBB.ProcessManager.Runtime/DependencyInjectionExtensions.cs
@@ -1,20 +1,25 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using NBB.ProcessManager.Runtime.Persistence;
-using NBB.ProcessManager.Runtime.Timeouts;
-using System;
+﻿using MediatR;
+using Microsoft.Extensions.DependencyInjection;
 using NBB.Application.Effects;
 using NBB.Core.Effects;
 using NBB.Http.Effects;
 using NBB.Messaging.Effects;
+using NBB.ProcessManager.Definition;
 using NBB.ProcessManager.Definition.SideEffects;
+using NBB.ProcessManager.Runtime.Persistence;
 using NBB.ProcessManager.Runtime.SideEffectHandlers;
+using NBB.ProcessManager.Runtime.Timeouts;
+using System;
+using System.Reflection;
+using Unit = NBB.Core.Effects.Unit;
 
 namespace NBB.ProcessManager.Runtime
 {
     public static class DependencyInjectionExtensions
     {
-        public static void AddProcessManagerRuntime(this IServiceCollection services)
+        public static void AddProcessManager(this IServiceCollection services, params Assembly[] assemblies)
         {
+            services.AddProcessManagerDefinition(assemblies);
             services.AddSingleton<ProcessExecutionCoordinator>();
             services.AddSingleton<IInstanceDataRepository, InstanceDataRepository>();
             services.AddEffects();
@@ -22,9 +27,11 @@ namespace NBB.ProcessManager.Runtime
             services.AddMessagingEffects();
             services.AddHttpEffects();
             services.AddMediatorEffects();
+            services.AddScoped<INotificationHandler<TimeoutOccured>, TimeoutOccuredHandler>();
+            services.AddNotificationHandlers(typeof(ProcessManagerNotificationHandler<,,>));
         }
 
-        public static void AddTimeoutEffects(this IServiceCollection services)
+        private static void AddTimeoutEffects(this IServiceCollection services)
         {
             services.AddHostedService<TimeoutsService>();
             services.AddSingleton<TimeoutsManager>();
